Protocol in progress; SenderKeyDistributionMessage and SenderKeyMessage complete

+++ Axolotl/Groups/GroupCipher.cs
@@ -16,34 +16,35 @@ namespace Axolotl.Groups
 		private ISenderKeyStore _senderKeyStore;
 		private SenderKeyName _senderKeyId;
 
-		public GroupCipher (ISenderKeyStore senderKeyStore, SenderKeyName senderKeyId)
+		public GroupCipher(ISenderKeyStore senderKeyStore, SenderKeyName senderKeyId)
 		{
 			_senderKeyStore = senderKeyStore;
 			_senderKeyId = senderKeyId;
 		}
 
-		public byte[] Encrypt(byte[] paddedPlaintext) 
+		public byte[] Encrypt(byte[] paddedPlaintext)
 		{
-			lock(LOCK) {
+			lock(LOCK)
+			{
 				// TODO
-	//			try {
-					var record         = _senderKeyStore.LoadSenderKey(_senderKeyId);
-					var senderKeyState = record.GetSenderKeyState();
-					var senderKey      = senderKeyState.SenderChainKey.GetSenderMessageKey();
-					var ciphertext     = GetCipherText(senderKey.Iv, senderKey.CipherKey, paddedPlaintext);
+				//			try {
+				var record = _senderKeyStore.LoadSenderKey(_senderKeyId);
+				var senderKeyState = record.GetSenderKeyState();
+				var senderKey = senderKeyState.SenderChainKey.GetSenderMessageKey();
+				var ciphertext = GetCipherText(senderKey.Iv, senderKey.CipherKey, paddedPlaintext);
 
-					var senderKeyMessage = new SenderKeyMessage(senderKeyState.KeyId,
-					                                            senderKey.Iteration,
-					                                            ciphertext,
-					                                            senderKeyState.SigningKeyPrivate);
+				var senderKeyMessage = new SenderKeyMessage((int)senderKeyState.KeyId,
+					                       senderKey.Iteration,
+					                       ciphertext,
+					                       senderKeyState.SigningKeyPrivate);
 
-					senderKeyState.SenderChainKey = senderKeyState.SenderChainKey.GetNext();
-					_senderKeyStore.StoreSenderKey(_senderKeyId, record);
+				senderKeyState.SenderChainKey = senderKeyState.SenderChainKey.GetNext();
+				_senderKeyStore.StoreSenderKey(_senderKeyId, record);
 
-					return senderKeyMessage.Serialize();
-	//			} catch (InvalidKeyIdException e) {
-	//				throw new NoSessionException(e);
-	//			}
+				return senderKeyMessage.Serialize();
+				//			} catch (InvalidKeyIdException e) {
+				//				throw new NoSessionException(e);
+				//			}
 			}
 		}
 
@@ -55,15 +56,17 @@ namespace Axolotl.Groups
 		public byte[] Decrypt(byte[] senderKeyMessageBytes, IDecryprionCallback callback)
 		{
 			//TODO Try/catch and stuff
-			lock (LOCK) {
+			lock(LOCK)
+			{
 				var record = _senderKeyStore.LoadSenderKey(_senderKeyId);
 
-				if (record.IsEmpty) {
+				if(record.IsEmpty)
+				{
 					throw new Exception("No sender key for: " + _senderKeyId);
 				}
 
 				var senderKeyMessage = new SenderKeyMessage(senderKeyMessageBytes);
-				var senderKeyState   = record.GetSenderKeyState(senderKeyMessage.KeyId);
+				var senderKeyState = record.GetSenderKeyState(senderKeyMessage.KeyId);
 
 				senderKeyMessage.VerifySignature(senderKeyState.SigningKeyPublic);
 
@@ -79,24 +82,30 @@ namespace Axolotl.Groups
 			}
 		}
 
-		private SenderMessageKey GetSenderKey (SenderKeyState senderKeyState, int iteration)
+		private SenderMessageKey GetSenderKey(SenderKeyState senderKeyState, int iteration)
 		{
 			var senderChainKey = senderKeyState.SenderChainKey;
 
-			if (senderChainKey.Iteration > iteration) {
-				if (senderKeyState.HasSenderMessageKey(iteration)) {
+			if(senderChainKey.Iteration > iteration)
+			{
+				if(senderKeyState.HasSenderMessageKey(iteration))
+				{
 					return senderKeyState.RemoveSenderMessageKey(iteration);
-				} else {
+				}
+				else
+				{
 					throw new Exception("Received message with old counter: " +
-					                                    senderChainKey.Iteration + " , " + iteration);
+					senderChainKey.Iteration + " , " + iteration);
 				}
 			}
 
-			if (senderChainKey.Iteration - iteration > 2000) {
+			if(senderChainKey.Iteration - iteration > 2000)
+			{
 				throw new Exception("Over 2000 messages into the future!");
 			}
 
-			while (senderChainKey.Iteration < iteration) {
+			while(senderChainKey.Iteration < iteration)
+			{
 				senderKeyState.AddSenderMessageKey(senderChainKey.GetSenderMessageKey());
 				senderChainKey = senderChainKey.GetNext();
 			}
@@ -105,52 +114,57 @@ namespace Axolotl.Groups
 			return senderChainKey.GetSenderMessageKey();
 		}
 
-		private byte[] GetPlainText (byte[] iv, byte[] key, byte[] ciphertext)
+		private byte[] GetPlainText(byte[] iv, byte[] key, byte[] ciphertext)
 		{
 			byte[] result; 
 
-			using (var rijAlg = Rijndael.Create()) {
+			using(var rijAlg = Rijndael.Create())
+			{
 				rijAlg.Key = key;
 				rijAlg.IV = iv;
 
-				var decryptor = rijAlg.CreateDecryptor (key, iv);
+				var decryptor = rijAlg.CreateDecryptor(key, iv);
 
-				using (var mStream = new MemoryStream())
-				using (var dcStream = new CryptoStream(mStream, decryptor, CryptoStreamMode.Read))
-				using (var sReader = new StreamReader(dcStream)) {
-					var stringText = sReader.ReadToEnd ();
-					result = Encoding.UTF8.GetBytes (stringText);
-				}
+				using(var mStream = new MemoryStream())
+					using(var dcStream = new CryptoStream(mStream, decryptor, CryptoStreamMode.Read))
+						using(var sReader = new StreamReader(dcStream))
+						{
+							var stringText = sReader.ReadToEnd();
+							result = Encoding.UTF8.GetBytes(stringText);
+						}
 			}
 
 			return result;
 		}
 
-		private byte[] GetCipherText (byte[] iv, byte[] cipherKey, byte[] paddedPlaintext)
+		private byte[] GetCipherText(byte[] iv, byte[] cipherKey, byte[] paddedPlaintext)
 		{
 			byte[] result;
 
-			using (var aes = SymmetricAlgorithm.Create()) {
+			using(var aes = SymmetricAlgorithm.Create())
+			{
 				aes.Mode = CipherMode.CBC;
 				aes.Key = cipherKey;
 				aes.IV = iv;
 				aes.Padding = PaddingMode.PKCS7;
-				var encryptor = aes.CreateEncryptor ();
+				var encryptor = aes.CreateEncryptor();
 
-				using (var mStream = new MemoryStream()) 
-				using (var cStream = new CryptoStream(mStream, encryptor, CryptoStreamMode.Write)) {
-					cStream.Write (paddedPlaintext, 0, paddedPlaintext.Length);
-					cStream.FlushFinalBlock ();
-					result = mStream.ToArray ();
-				}
+				using(var mStream = new MemoryStream())
+					using(var cStream = new CryptoStream(mStream, encryptor, CryptoStreamMode.Write))
+					{
+						cStream.Write(paddedPlaintext, 0, paddedPlaintext.Length);
+						cStream.FlushFinalBlock();
+						result = mStream.ToArray();
+					}
 			}
 
 			return result;
 		}
 
-		private class NullDecryptionCallback : IDecryprionCallback 
+		private class NullDecryptionCallback : IDecryprionCallback
 		{
-			public void HandlePlaintext(byte[] plaintext) {
+			public void HandlePlaintext(byte[] plaintext)
+			{
 			}
 		}
 	}
